SupplyGraph/Backend/ml-service/requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
networkx>=2.6.0
matplotlib>=3.5.0
seaborn>=0.11.0
flask>=2.0.0
flask-cors>=3.0.0
python-dotenv>=0.19.0
pymongo>=4.0.0 